# [BugFix]: Add ColName to MasterIndexScan for Filter PushDown