--- a/test/distributed/cases/ddl/secondary_index_master.result
+++ b/test/distributed/cases/ddl/secondary_index_master.result
@@ -1,4 +1,3 @@
-SET experimental_master_index = 1;
 drop table if exists t1;
 create table t1(a varchar(30), b varchar(30), c varchar(30) primary key);
 insert into t1 values("Congress","Lane", "1");
@@ -297,4 +296,3 @@ Nightingale    Lane    3
 select * from t1 where a between "Congress" and "Nightingale" and b="Lane" and c between "1" and "3" limit 1;
 a    b    c
 Congress    Lane    1
-SET experimental_master_index = 0;