### **User description**
Remove Master Index Experimental Flag

## What type of PR is this?

- [ ] API-change
- [x] BUG
- [ ] Improvement
- [ ] Documentation
- [ ] Feature
- [ ] Test and CI
- [ ] Code Refactoring

## Which issue(s) this PR fixes:

issue # https://github.com/matrixorigin/matrixone/issues/14876

## What this PR does / why we need it:

- Adding ColName to MasterIndex Optimizer Plan for Filter Pushdown.
- With this change, master index performance is now reasonable. Hence removing the Experimental Flag. 

1 Filter Query QPS
- No index: 500
- 1 Master: 2820
- 100 Secondary: 2958

<details>
<summary> Query Plan </summary>

```sql
-- Master Index

mysql> explain analyze SELECT tbl.a100  FROM tbl  WHERE tbl.a48 = 'b92k7dWP5t';
+-----------------------------------------------------------------------------------------------------------------------------------------------------+
| AP QUERY PLAN ON MULTICN(10 core)                                                                                                                   |
+-----------------------------------------------------------------------------------------------------------------------------------------------------+
| Project                                                                                                                                             |
|   Analyze: timeConsumed=0ms waitTime=4ms inputRows=2 outputRows=1 InputSize=48bytes OutputSize=24bytes MemorySize=0bytes                            |
|   ->  Join                                                                                                                                          |
|         Analyze: timeConsumed=1ms waitTime=29ms inputRows=2 outputRows=1 InputSize=48bytes OutputSize=24bytes MemorySize=0bytes                     |
|         Join Type: INDEX                                                                                                                            |
|         Join Cond: (tbl.a100 = __mo_index_secondary_01900572-3257-76e5-9a7d-e4eaa2a28f17.__mo_index_pri_col)                                        |
|         Runtime Filter Build: #[-1,0]                                                                                                               |
|         ->  Table Scan on a.tbl [ForceOneCN]                                                                                                        |
|               Analyze: timeConsumed=0ms waitTime=0ms inputBlocks=1 inputRows=1 outputRows=1 InputSize=48bytes OutputSize=24bytes MemorySize=50bytes |
|               Filter Cond: (tbl.a48 = 'b92k7dWP5t')                                                                                                 |
|               Block Filter Cond: (tbl.a48 = 'b92k7dWP5t')                                                                                           |
|               Runtime Filter Probe: tbl.a100                                                                                                        |
|         ->  Table Scan on a.__mo_index_secondary_01900572-3257-76e5-9a7d-e4eaa2a28f17 [ForceOneCN]                                                  |
|               Analyze: timeConsumed=1ms waitTime=0ms inputBlocks=1 inputRows=1 outputRows=1 InputSize=79bytes OutputSize=24bytes MemorySize=80bytes |
|               Filter Cond: prefix_eq(__mo_index_secondary_01900572-3257-76e5-9a7d-e4eaa2a28f17.__mo_index_idx_col, 'F47 Fb92k7dWP5t ')            |
|               Block Filter Cond: prefix_eq(__mo_index_secondary_01900572-3257-76e5-9a7d-e4eaa2a28f17.__mo_index_idx_col, 'F47 Fb92k7dWP5t ')      |
+-----------------------------------------------------------------------------------------------------------------------------------------------------+
16 rows in set (0.00 sec)


-- Secondary Index

mysql> explain analyze SELECT tbl.a100  FROM tbl  WHERE tbl.a48 = 'b92k7dWP5t';
+-----------------------------------------------------------------------------------------------------------------------------------------------------+
| TP QURERY PLAN                                                                                                                                      |
+-----------------------------------------------------------------------------------------------------------------------------------------------------+
| Project                                                                                                                                             |
|   Analyze: timeConsumed=0ms waitTime=0ms inputRows=2 outputRows=1 InputSize=48bytes OutputSize=24bytes MemorySize=0bytes                            |
|   ->  Join                                                                                                                                          |
|         Analyze: timeConsumed=0ms waitTime=1ms inputRows=1 outputRows=1 InputSize=24bytes OutputSize=24bytes MemorySize=0bytes                      |
|         Join Type: INDEX                                                                                                                            |
|         Join Cond: (tbl.a100 = __mo_index_secondary_01900577-f81d-7a7b-802c-b61a09a28067.__mo_index_pri_col)                                        |
|         Runtime Filter Build: #[-1,0]                                                                                                               |
|         ->  Table Scan on a.tbl [ForceOneCN]                                                                                                        |
|               Analyze: timeConsumed=0ms waitTime=0ms inputBlocks=1 inputRows=1 outputRows=1 InputSize=48bytes OutputSize=24bytes MemorySize=50bytes |
|               Filter Cond: (tbl.a48 = 'b92k7dWP5t')                                                                                                 |
|               Block Filter Cond: (tbl.a48 = 'b92k7dWP5t')                                                                                           |
|               Runtime Filter Probe: tbl.a100                                                                                                        |
|         ->  Table Scan on a.__mo_index_secondary_01900577-f81d-7a7b-802c-b61a09a28067 [ForceOneCN]                                                  |
|               Analyze: timeConsumed=0ms waitTime=0ms inputBlocks=1 inputRows=1 outputRows=1 InputSize=74bytes OutputSize=24bytes MemorySize=75bytes |
|               Filter Cond: prefix_eq(__mo_index_secondary_01900577-f81d-7a7b-802c-b61a09a28067.__mo_index_idx_col, 'Fb92k7dWP5t ')                 |
|               Block Filter Cond: prefix_eq(__mo_index_secondary_01900577-f81d-7a7b-802c-b61a09a28067.__mo_index_idx_col, 'Fb92k7dWP5t ')           |
+-----------------------------------------------------------------------------------------------------------------------------------------------------+
16 rows in set (0.00 sec)



```
</details>


2 Filter Query QPS
- No Index: 
- 1 Master: 1335 (right now we use both the filters using inner join)
- 100 Secondary: 1725 (we only make use of one secondary index table)

<details>
<summary> Query Plan </summary>


```sql
-- master index
mysql> explain analyze SELECT tbl.a100  FROM tbl  WHERE tbl.a89 = '40u4JSeGvz' AND tbl.a31 = '3X5ZOcJbol';
+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| AP QUERY PLAN ON MULTICN(10 core)                                                                                                                                                                            |
+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
| Project                                                                                                                                                                                                      |
|   Analyze: timeConsumed=0ms waitTime=36ms inputRows=3 outputRows=1 InputSize=72bytes OutputSize=24bytes MemorySize=0bytes                                                                                    |
|   ->  Join                                                                                                                                                                                                   |
|         Analyze: timeConsumed=11ms waitTime=167ms inputRows=2 outputRows=1 InputSize=48bytes OutputSize=24bytes MemorySize=0bytes                                                                            |
|         Join Type: INDEX                                                                                                                                                                                     |
|         Join Cond: (tbl.a100 = __mo_index_secondary_01900572-3257-76e5-9a7d-e4eaa2a28f17.__mo_index_pri_col)                                                                                                 |
|         Runtime Filter Build: #[-1,0]                                                                                                                                                                        |
|         ->  Table Scan on a.tbl [ForceOneCN]                                                                                                                                                                 |
|               Analyze: timeConsumed=0ms waitTime=0ms inputBlocks=1 inputRows=1 outputRows=1 InputSize=72bytes OutputSize=24bytes MemorySize=75bytes                                                          |
|               Filter Cond: (tbl.a89 = '40u4JSeGvz'), (tbl.a31 = '3X5ZOcJbol')                                                                                                                                |
|               Block Filter Cond: (tbl.a89 = '40u4JSeGvz'), (tbl.a31 = '3X5ZOcJbol')                                                                                                                          |
|               Runtime Filter Probe: tbl.a100                                                                                                                                                                 |
|         ->  Join                                                                                                                                                                                             |
|               Analyze: timeConsumed=11ms probe_time=[total=0ms,min=0ms,max=0ms,dop=10] build_time=[11ms] waitTime=145ms inputRows=2 outputRows=1 InputSize=48bytes OutputSize=24bytes MemorySize=361187bytes |
|               Join Type: INNER                                                                                                                                                                               |
|               Join Cond: (__mo_index_secondary_01900572-3257-76e5-9a7d-e4eaa2a28f17.__mo_index_pri_col = __mo_index_secondary_01900572-3257-76e5-9a7d-e4eaa2a28f17.__mo_index_pri_col)                       |
|               ->  Table Scan on a.__mo_index_secondary_01900572-3257-76e5-9a7d-e4eaa2a28f17 [ForceOneCN]                                                                                                     |
|                     Analyze: timeConsumed=5ms waitTime=0ms inputBlocks=1 inputRows=1 outputRows=1 InputSize=79bytes OutputSize=24bytes MemorySize=80bytes                                                    |
|                     Filter Cond: prefix_eq(__mo_index_secondary_01900572-3257-76e5-9a7d-e4eaa2a28f17.__mo_index_idx_col, 'F30 F3X5ZOcJbol ')                                                               |
|                     Block Filter Cond: prefix_eq(__mo_index_secondary_01900572-3257-76e5-9a7d-e4eaa2a28f17.__mo_index_idx_col, 'F30 F3X5ZOcJbol ')                                                         |
|               ->  Table Scan on a.__mo_index_secondary_01900572-3257-76e5-9a7d-e4eaa2a28f17 [ForceOneCN]                                                                                                     |
|                     Analyze: timeConsumed=11ms waitTime=0ms inputBlocks=1 inputRows=1 outputRows=1 InputSize=79bytes OutputSize=24bytes MemorySize=80bytes                                                   |
|                     Filter Cond: prefix_eq(__mo_index_secondary_01900572-3257-76e5-9a7d-e4eaa2a28f17.__mo_index_idx_col, 'F88 F40u4JSeGvz ')                                                               |
|                     Block Filter Cond: prefix_eq(__mo_index_secondary_01900572-3257-76e5-9a7d-e4eaa2a28f17.__mo_index_idx_col, 'F88 F40u4JSeGvz ')                                                         |
+--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------+
24 rows in set (0.03 sec)



-- secondary index

mysql> explain analyze SELECT tbl.a100  FROM tbl  WHERE tbl.a89 = '40u4JSeGvz' AND tbl.a31 = '3X5ZOcJbol';
+-----------------------------------------------------------------------------------------------------------------------------------------------------+
| TP QURERY PLAN                                                                                                                                      |
+-----------------------------------------------------------------------------------------------------------------------------------------------------+
| Project                                                                                                                                             |
|   Analyze: timeConsumed=0ms waitTime=0ms inputRows=2 outputRows=1 InputSize=48bytes OutputSize=24bytes MemorySize=0bytes                            |
|   ->  Join                                                                                                                                          |
|         Analyze: timeConsumed=0ms waitTime=1ms inputRows=1 outputRows=1 InputSize=24bytes OutputSize=24bytes MemorySize=0bytes                      |
|         Join Type: INDEX                                                                                                                            |
|         Join Cond: (tbl.a100 = __mo_index_secondary_01900568-4e08-7f3a-9fb9-755355944df6.__mo_index_pri_col)                                        |
|         Runtime Filter Build: #[-1,0]                                                                                                               |
|         ->  Table Scan on a.tbl [ForceOneCN]                                                                                                        |
|               Analyze: timeConsumed=0ms waitTime=0ms inputBlocks=1 inputRows=1 outputRows=1 InputSize=72bytes OutputSize=24bytes MemorySize=75bytes |
|               Filter Cond: (tbl.a89 = '40u4JSeGvz'), (tbl.a31 = '3X5ZOcJbol')                                                                       |
|               Block Filter Cond: (tbl.a89 = '40u4JSeGvz'), (tbl.a31 = '3X5ZOcJbol')                                                                 |
|               Runtime Filter Probe: tbl.a100                                                                                                        |
|         ->  Table Scan on a.__mo_index_secondary_01900568-4e08-7f3a-9fb9-755355944df6 [ForceOneCN]                                                  |
|               Analyze: timeConsumed=0ms waitTime=0ms inputBlocks=1 inputRows=1 outputRows=1 InputSize=74bytes OutputSize=24bytes MemorySize=75bytes |
|               Filter Cond: prefix_eq(__mo_index_secondary_01900568-4e08-7f3a-9fb9-755355944df6.__mo_index_idx_col, 'F3X5ZOcJbol ')                 |
|               Block Filter Cond: prefix_eq(__mo_index_secondary_01900568-4e08-7f3a-9fb9-755355944df6.__mo_index_idx_col, 'F3X5ZOcJbol ')           |
+-----------------------------------------------------------------------------------------------------------------------------------------------------+
16 rows in set (0.00 sec)
```
</details>


___

### **PR Type**
Bug fix, Enhancement


___

### **Description**
- Removed the `experimental_master_index` system variable and related checks.
- Added column name setting for `ColExpr` to be pushed down to the storage engine layer in the master index.
- Updated test cases and results to remove the setting of `experimental_master_index`.



___



### **Changes walkthrough** 📝
<table><thead><tr><th></th><th align="left">Relevant files</th></tr></thead><tbody><tr><td><strong>Enhancement
</strong></td><td><table>
<tr>
  <td>
    <details>
      <summary><strong>variables.go</strong><dd><code>Remove `experimental_master_index` system variable</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>
      
pkg/frontend/variables.go

- Removed the `experimental_master_index` system variable.



</details>
    

  </td>
  <td><a href="https://github.com/matrixorigin/matrixone/pull/16778/files#diff-07649adc637b0268b12ec915ec819f66db3a4e4375bfc77291116bfa06c4432f">+0/-8</a>&nbsp; &nbsp; &nbsp; </td>

</tr>                    

<tr>
  <td>
    <details>
      <summary><strong>ddl_index_algo.go</strong><dd><code>Remove checks for `experimental_master_index` flag</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>
      
pkg/sql/compile/ddl_index_algo.go

- Removed checks for the `experimental_master_index` flag.



</details>
    

  </td>
  <td><a href="https://github.com/matrixorigin/matrixone/pull/16778/files#diff-c6f4b272afb302e3b5d0d6460544032c03ac0ec988fd3cc8e89389ba657c0477">+0/-7</a>&nbsp; &nbsp; &nbsp; </td>

</tr>                    

<tr>
  <td>
    <details>
      <summary><strong>apply_indices_master.go</strong><dd><code>Add column name setting for `ColExpr` in master index</code>&nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>
      
pkg/sql/plan/apply_indices_master.go

<li>Added column name setting for <code>ColExpr</code> to be pushed down to the storage <br>engine layer.<br>


</details>
    

  </td>
  <td><a href="https://github.com/matrixorigin/matrixone/pull/16778/files#diff-54eef50579b1530b896371b1ac32b6e0a743cf473b18410ca47a0059b3facc50">+4/-0</a>&nbsp; &nbsp; &nbsp; </td>

</tr>                    
</table></td></tr><tr><td><strong>Tests
</strong></td><td><table>
<tr>
  <td>
    <details>
      <summary><strong>secondary_index_master.result</strong><dd><code>Remove `experimental_master_index` setting in test results</code></dd></summary>
<hr>
      
test/distributed/cases/ddl/secondary_index_master.result

- Removed setting of `experimental_master_index` in test results.



</details>
    

  </td>
  <td><a href="https://github.com/matrixorigin/matrixone/pull/16778/files#diff-e27c09b111ea1471cb244242b467a354cddf0a8591645bb0e3422d6c9de15362">+0/-2</a>&nbsp; &nbsp; &nbsp; </td>

</tr>                    

<tr>
  <td>
    <details>
      <summary><strong>secondary_index_master.sql</strong><dd><code>Remove `experimental_master_index` setting in test cases</code>&nbsp; </dd></summary>
<hr>
      
test/distributed/cases/ddl/secondary_index_master.sql

- Removed setting of `experimental_master_index` in test cases.



</details>
    

  </td>
  <td><a href="https://github.com/matrixorigin/matrixone/pull/16778/files#diff-f77028c1b7abd4aa18b06f21defcada2ee41b48de35585984ed273c7cbe970ca">+0/-2</a>&nbsp; &nbsp; &nbsp; </td>

</tr>                    

<tr>
  <td>
    <details>
…